+++ lib/vagrant-service-manager.rb
@@ -17,5 +17,7 @@ module Vagrant
 
     # Temporally load the extra capabilities files for Red Hat
     load(File.join(source_root, 'plugins/guests/redhat/plugin.rb'))
+    # Default I18n to load the en locale
+    I18n.load_path << File.expand_path("locales/en.yml", source_root)
   end
 end